requirements.txt
PyGithub==1.43.8
pandas==0.25.0
colorama==0.4.1
futures==3.1.1
retrying==1.3.3
boto3==1.9.215
botocore==1.12.215